requirement.txt
requests
pandas
scikit-learn
matplotlib
umap-learn
torch
transformers
fastapi 
uvicorn 
plotly